Reduced redmatter entity damage from ~100M to 2K
This is to fix issues with other mods that try to reduce and neglect the damage. In which the high damage would result in overflows, underflows, bugs, crashes, and odd behavior. This is a temp fix as a better fix would be to change the damage type and add logic to bypass armor.