Minor API change, Finished getting cluster missile working
Cluster missile now fires missiles correctly but will need more dev love later. Mainly to improve spawning to prevent glitching through walls. As well to improve graphics so spawning is not instant but time based. In which the missile should spawn from the core and move outwards before firing its engine.

Also with this I added a new flight mode called DEAD_AIM. Which should be used for missiles with no guidance at all. Additional to this I added code to set the flight path of the missile in a strait line based on a target point.