Fixed overlap between IInventory and IItemHandler
Seems both interfaces share the same method... but one will be changed into srg names on compile. This will cause crashes for mods trying to access the inventory.

To solve for this I removed the IItemHandler code from ExternalInventory. Then added a wrapper to go around the class to re-implement the support as a seperate object.