Added runtime check to redmatter to prevent thread stall
Temp fix, the code will need to be rewritten to prevent the issue. As the redmatter when it has no blocks can run so long as to stall the game out. This doesn't actually cause the game to stop working. Rather it gives the impression of a freeze. In order to fix this a simple runtime check of less than 30ms was added. This should prevent complete game failure.