Removed size scale on redmatter animation, Smoothed animation to be continious
Massive improvement on visuals for redmatter. First fix was removing size scale that caused a pulse that was painful too look at. After this animation time was increase from 200 to 4000, with a ramp up and down time. This time causes the beam count to increase and decrease over 3.3 mins. Creating a very smooth animation that when combined with the rotation changes looks very good.

Rotation change involves seperating out the time scale from rotation scale. It now rotations about 200 tick animation time. As well was changed from 90 degress to 360. This by itself results in a very smooth animation that is hard to notice changes.