Renamed position to location, Started work on replacing Explosion handlers with lambda factory
Rename is to avoid conflict with vanilla using the same field name in a super class.

Lambda factory will replace basic usage with the goal to completely phase out the handlers. The idea will be to register simple interfaces to handle creating the blast and displaying data. Reg of explosive to each type will be handled by a different system. This way we can free up as much bloat and refine the explosive system. Then once complete we can implement an API to register new explosive types.