Implemented simple blast factory interface
Idea is that we will build blasts using a factory method/expression that calls to interfaces. The interface will allow changing the properties of the blast without understanding how it works internally. As well it will prevent changing the blast after it has been created. With an event being implemented to allow overriding before creation.