Started on rebuilding redmatter as an entity
Goal with this rework is to move the redmatter logic to a custom entity. As it stands the redmatter is working around the blast system. A lot of custom logic was added to the blasts to let the redmatter function. Such as IMovableBlast and special logic to work around disabling yOffsets.

 The original idea for making the redmatter a blast instance was to de-duplicate work. However, this only resulted in a complicated system that doesn't improve the quality of functionality. The blast version caused the redmatter to be more abstracted then required. Making for more code and an overall more complicated system.