"Fix" ConcurrentModificationException in RadarMap
There is a lack of thread safety on the lists stored in
the chunk_to_entities HashMap in RadarMap.

When a chunk is unloaded, MissileEventHandler.chunkUnload calls
collectEntitiesInChunk and adds any missiles to the
MissileTrackerHandler for simulation.
It looks like somewhere else in the code we are modifying the list that
is being iterated in chunkUnload but I havn't found out where yet.

As a temporary fix I added synchronized() blocks around everywhere I
found modifies the list and I made getEntitiesInChunk return a copy of
the list.

So far I havn't had any crashes but i'm not convinced this is the best
solution

Issue mentioned in #337, #312, and #295