fix: duplication bug during transfer stack during chunk unload
This is in addition to the previous commit. As that should for the most part prevent the majority of players from exploiting the system. This just added to that protection by rejecting any transferring of inventory between player and tile. Its not likely going to stop all issues but should narrow down the options.